Fixed bug in Availabilities
Picking Sunday in the availability day returns a failure state since the
Sunday entry is at index 0 in the DAYS_OF_THE_WEEK array in the
availabilities controller. As such, it gets evaluated to false when
checking if the day data exist/is valid or not. Changed the validation
to include if the day data is int (0) in order to account for this.